projects/formation-ia-2026/projets/semaine-1-assistant-recherche/main.py
"""
Assistant de Recherche Intelligent
Point d'entrée CLI

Usage:
    python main.py
"""

import os
from anthropic import Anthropic
from dotenv import load_dotenv

# Charger les variables d'environnement
load_dotenv()

# Initialiser le client Claude
client = Anthropic(api_key=os.environ.get("ANTHROPIC_API_KEY"))


class ResearchAssistant:
    """Assistant de recherche intelligent"""

    def __init__(self):
        self.client = client
        self.history = []

    def research(self, question: str) -> str:
        """
        Effectue une recherche et retourne la synthèse

        Args:
            question: Question de l'utilisateur

        Returns:
            Synthèse de la recherche
        """
        # TODO: Implémenter la logique de recherche
        # 1. Décomposer la question si nécessaire
        # 2. Chercher des sources (Wikipedia, etc.)
        # 3. Synthétiser les résultats
        pass

    def decompose_question(self, question: str) -> list:
        """
        Décompose une question complexe en sous-questions

        Args:
            question: Question complexe

        Returns:
            Liste de sous-questions
        """
        # TODO: Utiliser Claude pour décomposer la question
        pass

    def search_sources(self, query: str) -> list:
        """
        Cherche des informations dans différentes sources

        Args:
            query: Requête de recherche

        Returns:
            Liste de résultats
        """
        # TODO: Implémenter la recherche multi-sources
        pass

    def synthesize(self, sources: list, question: str) -> str:
        """
        Synthétise les informations trouvées

        Args:
            sources: Liste de sources d'information
            question: Question originale

        Returns:
            Synthèse formatée
        """
        # TODO: Utiliser Claude pour synthétiser
        pass


def main():
    """Point d'entrée principal"""

    print("🔍 Assistant de Recherche Intelligent")
    print("=" * 50)
    print("Tapez 'quit' pour quitter\n")

    assistant = ResearchAssistant()

    while True:
        try:
            question = input("\n📝 Votre question : ").strip()

            if not question:
                continue

            if question.lower() in ['quit', 'exit', 'q']:
                print("\n👋 Au revoir !")
                break

            print("\n🔍 Recherche en cours...\n")

            answer = assistant.research(question)

            print(f"📚 Réponse :\n{answer}\n")
            print("-" * 50)

        except KeyboardInterrupt:
            print("\n\n👋 Au revoir !")
            break
        except Exception as e:
            print(f"\n❌ Erreur : {e}\n")


if __name__ == "__main__":
    main()
