projects/formation-ia-2026/projets/semaine-1-assistant-recherche/requirements.txt
anthropic>=0.18.0
python-dotenv>=1.0.0
wikipedia>=1.4.0
requests>=2.31.0
fastapi>=0.109.0
uvicorn>=0.27.0
pydantic>=2.5.0
